refactor helper view fns into own file, also data constant
